Python/app_close/requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
enum34==1.1.6
ipaddress==1.0.22
mysql-connector==2.2.9
pycparser==2.19
six==1.12.0
